dev/4_microservices/lists/requirements/requirements-back.txt
grpc-interceptor ~= 0.11.0
grpcio-tools ~= 1.30
google-cloud-bigquery ~= 3.0.1